fix: :bug: Ensure grindKey grinds to the correct value when keySeed has leading zeroes
Since keySeed is a big number, if it is smaller than 2^252, when it gets converted back
to a hex string, the string has any leading zeroes stripped.

This means the hash grind loop arriving at an incorrect result (the bug)
The fix was to ensure that the input to the hash loop was exactly 64
characters long, to represent a keySeed value range of [0, 2^256), which
ensured the correct number of leading zeroes would be used.