TD-503 Revert version number in package.json and CHANGELOG.md (#355)
Revert version number in package.json and CHANGELOG.md

The version numbers in both package.json and CHANGELOG.md were mistakenly updated to 2.0.3. This commit reverts the version back to 2.0.2 in package.json, and changes the version title in the CHANGELOG.md to "Unreleased". The updates made were premature and the version number should only be changed upon official release.