TD-803 Update version to 2.1.0 in changelog and package.json (#356)
Update version to 2.1.0 in changelog and package.json

This commit changes the version of the project from 2.0.2 to 2.1.0 in both the package.json and the CHANGELOG.md files. The 'unreleased' tag in the changelog is also updated to reflect the new version. This change is made to mark the completion of new features and enhancements for the 2.1.0 release.